+++ Edadil_MicroService/src/main/java/com/example/edadil_microservice/repository/CompanyRepository.java
@@ -7,13 +7,13 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
+public interface CompanyRepository extends JpaRepository<Company, Long> {
 
 
     @Query("SELECT DISTINCT sp.shop.nameOfCompany FROM ShopProduct sp WHERE sp.product.firm.id = :firmId")
-    List<Company> findCompaniesByFirmId(Integer firmId);
+    List<Company> findCompaniesByFirmId(Long firmId);
 
     @Query("SELECT DISTINCT sp.shop.nameOfCompany FROM ShopProduct sp WHERE sp.product.firm.id = :firmId AND sp.shop.nameOfCompany.id = :companyId")
-    Optional<Company> findByFirmIdAndCompanyId(Integer firmId, Integer companyId);
+    Optional<Company> findByFirmIdAndCompanyId(Long firmId, Long companyId);
 
 }
