+++ Edadil_MicroService/src/main/java/com/example/edadil_microservice/repository/FirmRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface FirmRepository extends JpaRepository<Firm, Integer> {
+public interface FirmRepository extends JpaRepository<Firm, Long> {
 
 
     @Query("SELECT f FROM Firm f JOIN FETCH f.products WHERE f.id = :firmId")
-    Optional<Firm> findById(Integer firmId);
+    Optional<Firm> findById(Long firmId);
 }
