+++ Edadil_MicroService/src/main/java/com/example/edadil_microservice/repository/ShopProductRepository.java
@@ -7,19 +7,19 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ShopProductRepository extends JpaRepository<ShopProduct, Integer> {
+public interface ShopProductRepository extends JpaRepository<ShopProduct, Long> {
     @EntityGraph(attributePaths = {"shop", "shop.nameOfCompany", "product", "product.firm"})
-    List<ShopProduct> findShopProductsByProductId(Integer productId);
+    List<ShopProduct> findShopProductsByProductId(Long productId);
 
     @Query("SELECT DISTINCT sp FROM ShopProduct sp JOIN FETCH sp.shop s JOIN FETCH s.nameOfCompany c JOIN FETCH sp.product p JOIN FETCH p.firm f WHERE f.id = :firmId AND c.id = :companyId")
-    List<ShopProduct> findShopProductsByFirmIdAndCompanyId(Integer firmId, Integer companyId);
+    List<ShopProduct> findShopProductsByFirmIdAndCompanyId(Long firmId, Long companyId);
 
     @Query("SELECT sp FROM ShopProduct sp WHERE sp.shop.nameOfCompany.id = :companyId AND sp.shop.city = :city AND sp.shop.id = :shopId")
-    List<ShopProduct> findShopProductsByCompanyIdAndCityAndShopId(Integer companyId, String city, Integer shopId);
+    List<ShopProduct> findShopProductsByCompanyIdAndCityAndShopId(Long companyId, String city, Long shopId);
 
     @Query("SELECT sp FROM ShopProduct sp JOIN FETCH sp.shop s JOIN FETCH s.nameOfCompany c JOIN FETCH sp.product p JOIN FETCH p.firm f WHERE f.id = :firmId AND c.id = :companyId AND s.id = :shopId")
-    List<ShopProduct> findProductsInShopByFirmAndCompany(Integer firmId, Integer companyId, Integer shopId);
+    List<ShopProduct> findProductsInShopByFirmAndCompany(Long firmId, Long companyId, Long shopId);
 
     @Query("SELECT sp FROM ShopProduct sp WHERE sp.shop.id = :shopId")
-    List<ShopProduct> findByShopId(Integer shopId);
+    List<ShopProduct> findByShopId(Long shopId);
 }
