+++ Edadil_MicroService/src/main/java/com/example/edadil_microservice/service/company/CompanyServiceImpl.java
@@ -31,36 +31,36 @@ public class CompanyServiceImpl implements CompanyService {
     }
 
     @Override
-    public CompanyDto findCompanyById(Integer companyId) {
+    public CompanyDto findCompanyById(Long companyId) {
         log.info("Fetching company with ID: {}", companyId);
         Company company = getCompanyById(companyId);
         return companyMapper.toDto(company);
     }
 
     @Override
-    public List<CompanyDto> findCompaniesSellingFirmProducts(Integer firmId) {
+    public List<CompanyDto> findCompaniesSellingFirmProducts(Long firmId) {
         log.info("Fetching companies selling products for firm with ID: {}", firmId);
         List<Company> companies = getCompaniesByFirmId(firmId);
         return companyMapper.toDtoList(companies);
     }
 
     @Override
-    public CompanyDto findCompanySellingFirmProductsById(Integer firmId, Integer companyId) {
+    public CompanyDto findCompanySellingFirmProductsById(Long firmId, Long companyId) {
         log.info("Fetching company with ID: {} selling products for firm with ID: {}", companyId, firmId);
         return companyMapper.toDto(getCompanyByFirmAndCompanyId(firmId, companyId));
     }
 
-    private Company getCompanyByFirmAndCompanyId(Integer firmId, Integer companyId) {
+    private Company getCompanyByFirmAndCompanyId(Long firmId, Long companyId) {
         return companyRepository.findByFirmIdAndCompanyId(firmId, companyId)
                 .orElseThrow(() -> new EntityNotFoundException("Company not found with firm ID: " + firmId + " and company ID: " + companyId));
     }
 
-    private Company getCompanyById(Integer companyId) {
+    private Company getCompanyById(Long companyId) {
         return companyRepository.findById(companyId)
                 .orElseThrow(() -> new EntityNotFoundException("Company not found with ID: " + companyId));
     }
 
-    private List<Company> getCompaniesByFirmId(Integer firmId) {
+    private List<Company> getCompaniesByFirmId(Long firmId) {
         List<Company> companies = companyRepository.findCompaniesByFirmId(firmId);
         if (companies.isEmpty()) {
             throw new NoCompaniesSellingFirmProductsException("No companies found selling products for firm ID: " + firmId);
