+++ Edadil_MicroService/src/main/java/com/example/edadil_microservice/service/firm/FirmService.java
@@ -8,6 +8,6 @@ public interface FirmService {
 
     List<FirmDto> findAllFirms();
 
-    FirmDto findFirmById(Integer firmId);
+    FirmDto findFirmById(Long firmId);
 
 }
